posix: allow setting statx' type of synchronization
statx(2) exposes 3 flags to set the sort of synchronization the
kernel will do on a remote filesystem:
 - AT_STATX_SYNC_AS_STAT
 - AT_STATX_FORCE_SYNC
 - AT_STATX_DONT_SYNC

This patch adds the RPBO_STATX_SYNC_TYPE option that takes one of
these flags and passes it down to the iterators returned by the
filter fsentries operation. The iterators then use it (if the
statx system call is available).

The default value for this option is AT_STATX_SYNC_AS_STAT

AT_STATX_FORCE_SYNC is not supported on platforms that do not provide
the statx() system call as metadata is not guaranteed to be synced.

Change-Id: I29750e7c1fe625ef751605356e23b8c00bb1ebe0
Signed-off-by: Quentin Bouget <quentin.bouget@cea.fr>