API: handle uninterpretable backend errors consistently
With libmongoc (and probably other database drivers), every database
operation can fail for a number of reasons. There are so many of those
reasons that it would be unreasonable for a single developper to
handle and/or document them all in a single commit because:
 - it would take a very long time;
 - it would not be relying on a stable API;
 - it would be almost impossible to test and verify.

Which is why we implement the following workaround:

A backend may fail to correctly interprate why one of its method
failed and must in those cases set errno to `RBH_BACKEND_ERROR` and
place a meaningful message in the buffer `rbh_backend_error`.
Users are then expected to report any occurence of such errors back
to backend maintainers which may then focus on better handling those
errors (that is set errno to a more meaningful value and document the
root cause of the error).

This means backend operations failing with errno set to
`RBH_BACKEND_ERROR` cannot be considered part of robinhood's API and
application writers should not rely on them (they should only prepare
their application to handle such an error). It follows that messages
stored in `rbh_backend_error` are not part of robinhood's API either.

Ideally no backend should use this mechanism.

Change-Id: I9d6c2a0eb9095601f445f2b6d272d98469e0666c
Signed-off-by: Quentin Bouget <quentin.bouget@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/cea-hpc/robinhood/+/471514
Reviewed-by: Dominique Martinet <asmadeus@codewreck.org>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>