fsentry: symlink must be the first field set
When rbh_fsentry_new() allocates a struct rbh_fsentry it asks for
enough memory to store a struct rbh_fsentry but also the data that
fsentry will point at. This allows easy memory management for the
caller: the returned pointer and all the data its point at can be
freed in a single call to free(3).

Since struct rbh_fsentry now ends with a flexible array (symlink),
when appending the data the structure will point at, one must be
careful to place that data _after_ the content of symlink.

To this end, this patch adds a comment so future maintainers and
contributors may not trip on this little implementation detail.

Change-Id: I5b2f92482c0dcac146e0f7c72a4aec6d2790f8c0
Signed-off-by: Quentin Bouget <quentin.bouget@cea.fr>