fix: memory leak in the posix iterator
While traversing a file system in the posix backend, we use
memoization to avoid recomputing a directory's ID for all its child
entries. The memory associated with that ID is freed when the
directories are processed in postorder (when all its children have
been processed).

When the iterator is destroyed before it is exhausted, there may still
be memoized IDs.

posix_iter_destroy() now makes sure these IDs are freed.

Reported-by: address-sanitizer
Change-Id: Iaa65ccd86f57a893d28415d1374c0a56132325d0
Signed-off-by: Quentin Bouget <quentin.bouget@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/cea-hpc/robinhood/+/479239
Reviewed-by: Dominique Martinet <asmadeus@codewreck.org>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>