fix(posix)!: unset the parent ID of a branch's root
The posix backend used to set the parent ID of a branch's root to the
branch's root own ID (ID == parent_ID). This is not correct if we
consider that a branch's root has a parent; it is also incorrect given
RobinHood's conventions:

  root.parent_id.size == 0 && root.name == ""

For the posix backend, the parent ID of a given entry is not easily
accessible. Instead of crippling performance (ie. traversing the whole
filesystem looking for the entry and deducing its parent ID), or
relying on FS-specific features (eg. llapi_fd2parent() for Lustre), we
choose to leave a branch's root parent ID unset for the posix backend.

Change-Id: I38b6c7df04f858a80cde2add780b572bd9e39320
Signed-off-by: Quentin Bouget <quentin.bouget@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/cea-hpc/robinhood/+/479240
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>