mongo: implement null filter support for servers < 3.6
Null filters always evaluate to `true`, their negation always
evaluates to false.

In MongoDB, a filter that always evaluates to `true` is the empty
filter (ie. {}). A filter that always evaluates to `false` can be
built using either {$expr: false} (it is the preferred method) or
{$where: false}.

Yet, servers before version 3.6 do not support the "$expr" operator,
and "$where" is not allowed in aggregation pipelines.

This patch chooses to implement filters that always evaluate to false
as: {_id: {$exists: false}}, since every document in a mongodb
database can be supposed to have an "_id" field.

Do note that documents without an "_id" field can technically be
inserted in a mongo collection (though it requires configuring the
collection that way). But RobinHood neither needs nor uses this
feature.

Signed-off-by: Quentin Bouget <quentin.bouget@cea.fr>
Change-Id: I596bd2fb955c25045bc921ee241b8d8541db31c5
Reviewed-on: https://review.gerrithub.io/c/cea-hpc/robinhood/+/482040
Reviewed-by: Dominique Martinet <asmadeus@codewreck.org>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>