mongo: fix binary handling when size == 0
MongoDB does not natively handle empty binary values. The accepted
workaround is to store "null"s instead.

This works pretty nicely, until you learn that {field: null} in a
mongo filter means: "field" is null, or "field" does not exist.
The mongo documentation suggests that to really check that a given
field is set to null, one should use: {field: {$type: 10}} (where
10 is MONGO_TYPE_NULL).

ref: https://docs.mongodb.com/manual/tutorial/query-for-null-fields/

Unfortunetaly, things are not quite so simple. RobinHood relies on
upserts to set fields to null. And:

db.entries.update({ns.parent: {$type: 10}}, {...}, {upsert: true})

will not actually set "ns.parent" to null in the case of an upsert.
To do it, one must instead run:

db.entries.update({ns.parent: {$eq: null, $type: 10}}, {...},
                  {upsert: true})

Note that adding {$type: 10} is only needed in filters, not in the
actual update documents (ie. "{...}" in the examples above).

Signed-off-by: Quentin Bouget <quentin.bouget@cea.fr>
Change-Id: Ic9aed5b50336467646290dbcd129891a9428c350
Reviewed-on: https://review.gerrithub.io/c/cea-hpc/robinhood/+/483586
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>