fix segfault when parsing a URI fails
Before parsing URIs, we register an exit handler to destroy any
struct rbh_backend we might instantiate.

Prior to this patch, the handler freed N backends, where N is the
number of URIs on the command line, whether those backends were
actually instantiated or not. Now, it frees M backends, where M is
the number of instantiated backends.

(Of course, after all the URIs on the command line are successfully
parsed, M == N)

Signed-off-by: Quentin Bouget <quentin.bouget@cea.fr>
Change-Id: I6ffa683661b0ed74d73dd428c45c7004c719e9e6
Reviewed-on: https://review.gerrithub.io/c/cea-hpc/rbh-find/+/483587
Reviewed-by: Dominique Martinet <asmadeus@codewreck.org>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>