mongo: remove workaround for empty binary values
libbson (>= 1.16) added support for empty binary values.

Before this patch, empty binary values were stored as null values.

This did not work as well as one could hope, as querying for null
fields in MongoDB is a bit tricky. [0] One of the most problematic
corner case is:

  SELECT * WHERE field IN [NULL, X, Y, Z]

Which I believe has to be converted to:

  SELECT * WHERE field IN [X, Y, Z] OR field IS NULL

Before it can be correctly translated to a mongo query that does the
right thing:

  {$or: [{field: {$in: [X, Y, Z]}}, {field: {$type: 10}}]}

[0] https://docs.mongodb.com/manual/tutorial/query-for-null-fields/

Signed-off-by: Quentin Bouget <quentin.bouget@cea.fr>
Change-Id: I0a26f63ded1f2855d8f1827a0ff87b2774d278be
Reviewed-on: https://review.gerrithub.io/c/cea-hpc/robinhood/+/493551
Reviewed-by: Dominique Martinet <asmadeus@codewreck.org>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>