utils: support URIs with paths for posix backends
Posix backends do not support any sort of filtering. But the generic
implementation of rbh_backend_from_uri() needs it.

This patch special cases posix backend URIs with paths so that they
can be made to work as well.

Note that posix backends require CAP_DAC_READ_SEARCH to branch(), and
this patch does not change that.

Signed-off-by: Quentin Bouget <quentin.bouget@cea.fr>
Change-Id: I481bb27683474a71fa4f84b4e09b65673600eb49
Reviewed-on: https://review.gerrithub.io/c/cea-hpc/librobinhood/+/506317
Reviewed-by: Dominique Martinet <asmadeus@codewreck.org>