itertools: remove rbh_mut_iter_chunkify()
The implementation could lead to memory leaks when sub-iterators were
freed before they were exhausted.

To detect if the original iterator is exhausted, a chunkify iterator
will try to pop an element. If there is one, it will yield a new chunk
that has a reference on that element. For mutable iterators, that
elements may have been malloc-ed and may need to be freed, but it may
also not be free-able. Who knows? Not me.

If this function is ever needed again, it should either not use the
"peeking" approach (eg. have chunks report to the chunkify iterator
that they exhausted the original iterator); or it should include a
boolean `freeable` in its prototype that tells the chunks wether or
not to free their first element if it was never actually yielded.

Signed-off-by: Quentin Bouget <quentin.bouget@cea.fr>
Change-Id: I7382206eb6c4662394b439db6a77685b091cd83a
Reviewed-on: https://review.gerrithub.io/c/cea-hpc/librobinhood/+/508349
Reviewed-by: Dominique Martinet <asmadeus@codewreck.org>