mongo: add safe check on the state of mongo cursor
This patch solves an issue triggered when using rbh-sync with
mongo as the source on a number of entries not divisible by
RBH_ITER_CHUNK_SIZE. In this case, when all entries are synced,
the mongoc cursor is done, but mongoc_cursor_next() is called
again, triggering an error.

However, this error is given when calling that function on a
completed or failed cursor, but mongoc doesn't allow us to
differentiate the two. Therefore, we check before calling
mongoc_cursor_next() that the cursor is still valid.

Change-Id: Ife7e2b6a85ede74ac59d1ac9bd4a6a3a0a2f486d
Signed-off-by: Yoann Valeri <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/cea-hpc/librobinhood/+/518333
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Quentin Bouget <ypsah@devyard.org>
Tested-by: Quentin Bouget <ypsah@devyard.org>