headers: use quotes rather than angle brackets for project headers
The meaning of each depends on the compiler [0], but for gcc [1] and
clang, this means that headers will be looked for in the local directory
first and then in the system directories.

This choice makes the most sense for robinhood.h and this commit mostly
brings consistency across the project.

[0] https://web.archive.org/web/20181230041359if_/http://www.open-std.org/jtc1/sc22/wg14/www/abq/c17_updated_proposed_fdis.pdf
    section 6.10.2
[1] https://gcc.gnu.org/onlinedocs/cpp/Include-Syntax.html

Signed-off-by: Quentin Bouget <ypsah@devyard.org>
Change-Id: I1c9651153a6e23cd3c96831686423b5f4929ece4
Reviewed-on: https://review.gerrithub.io/c/cea-hpc/librobinhood/+/523041
Reviewed-by: Dominique Martinet <asmadeus@codewreck.org>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>