find: fix incorrect index manipulation for action tokens
If an action were to be specified, we would consume the action
token and increase the index to get the potential file where
the result should be printed. But in the find function, we
would increase the index again to have it ready for the next
command line token.

As such, we would go over the number of command line arguments
if the action token and its parameter were the last one, and
trigger a wrong error, or we would skip the next token of the
command line.

Change-Id: I4ad1cf0465b61042244cc452622e0849c8076373
Signed-off-by: Yoann Valeri <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/cea-hpc/rbh-find/+/536737
Reviewed-by: Dominique Martinet <asmadeus@codewreck.org>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-by: Quentin Bouget <ypsah@devyard.org>
Tested-by: Quentin Bouget <ypsah@devyard.org>