posix: fix error on symlinks following
Since retrieving xattrs on files requires them to be
opened without the O_PATH flag, we end up with 2 issues:
performance (because of the filesystem traversal and
locking) and symlinks opening (as we left O_NOFOLLOW,
openning a symlink ends up returning an error).

To solve both issues, we use O_PATH when openning, and
instead of using the file descriptor returned, we
use the "/proc/self/fd" file to retrieve the xattrs
of the examined entry.

However, this opens up another issue because the file
is required to be properly opened by Lustre to retrieve
specific information regarding the layout or the MDT on
which the file is.

So we decided to open the file with O_NOFOLLOW, and if it
fails because the file is a symlink, reopen it with
O_PATH, and remove the retrieval of layout and MDT info
for symlinks.

Change-Id: I983be35f20ad44610e2ac9ab9cb4c532d78800d2
Signed-off-by: Yoann Valeri <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/cea-hpc/librobinhood/+/540363
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>