lustre: fix CREATE event parsing
The events were badly generated, putting for instance name and parent
information in the xattrs field of an UPSERT event. Now, the correct
fsevent type is used (ie. a LINK instead of an UPSERT for the inode
creation).

The CLOSE and MKDIR events are removed and will be fixed in future
patches.

Change-Id: Ia9253506e6024aa4bb1bd180c2b911299bd9aca3
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/cea-hpc/rbh-fsevents/+/546281
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>
Tested-by: VALERI Yoann <yoann.valeri@cea.fr>