filter: update hsm-state filtering
This patch updates the way the hsm-state 'none' is
filtered following the patch on librobinhood:
I1c941db5db1eeced644fb36c3cc0a802dd75dc8c.

Since HSM info are no longer kept in the DB if a
regular file is not archived, to keep the filtering
on the status 'none' means we have to check if the
field exists in the DB or not, which in turn means
we have to differentiate between regular files that
are not archived and the rest because both do not
have any hsm-state anymore.

Change-Id: I34b5dcc92356d0b6b70e3ff90297407df2d03196
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/cea-hpc/rbh-find-lustre/+/549575
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>