retention: add infinite retention data
This patch adds an infinite retention mechanism for files that will
eventually expire but not with a clear date.

To set it up, simply use the same extended attribute as before, but
instead of a number, give the value "inf". It will set up the expiration
date to the maximal unsigned int value.

Change-Id: I429bd64594692dc43f76108ff2b3f9093316e9f8
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/cea-hpc/librobinhood/+/552263
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>