cli: consider MOUNTPOINT as a RobinHood URI
As MOUNTPOINT is bound to a backend for attributes retrieval, it is
lighter to give it as a RobinHood URI instead of multiple options (one
for the kind of backend and another one for the mount point). It is also
simpler to invoke a backend given its URI, as shown in the next patch.

Change-Id: I76812387abb4f2cb788501bbf768b3701cc6b25c
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/cea-hpc/rbh-fsevents/+/552357
Reviewed-by: Dominique Martinet <asmadeus@codewreck.org>
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>