lustre: verify magic number macro availability
The two macros LOV_USER_MAGIC_SEL and LOV_USER_MAGIC_FOREIGN are
available in Lustre 2.15, but not in Lustre 2.12, leading to compilation
failure on certain systems. This patch adds a preliminary check at build
to see if those macros are defined, and if so, manage them in the Lustre
backend.

Change-Id: I005533225cff37016dcf047a6363474843aa8ef1
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/cea-hpc/librobinhood/+/552484
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>