tests: mutualize call to rbh-fsevents in a function
This patch adds a new function to test_utils.sh that is simply calls
rbh-fsevents, and modify every call to rbh-fsevents to use this function
instead. This will make it easier to modify when the source feed is
given as an URI.

Also reorder the functions in test_utils.sh to outline specific sections
of the file.

Change-Id: Ic0158492f87cfe7132df2836cedba933215ce56d
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/cea-hpc/rbh-fsevents/+/553441
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>