posix: fix memory leak inside posix_iter_next
When the global variables `pairs`, `values` and `xattrs` where made
local, they were set to NULL on each call to fsentry_from_ftsent. Which
means that the sstacks and the array where allocated on each call but
never freed. This patch reverts this change by making them global again.

To make memory leak easier to debug, destructors are added to cleanup
global variables that are allocated only once:

- `names` in getxattrs
- `handle` in id_from_fd
- `ns_pairs`, `ns_values`, `xattrs`, `values` and `pairs` in fsentry_from_ftsent

Change-Id: Id0dd31f3d26cddc9ea6ebf22acbeb07a88266529
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/cea-hpc/librobinhood/+/553381
Tested-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>