lustre: correctly retrieve lustre information
This patch corrects the retrieval of Lustre information for FLRW,
LAYOUT, RESYNC, MKDIR and MKNOD events, which were not retrieved.

Also update the corresponding tests to not check that Lustre values have
changed. More specific Lustre tests are added in a next patch.

Finally, add a O_NONBLOCK flag to the open for Lustre information
retrieval, without which using FIFOs would timeout.

Change-Id: I93ebc9bf126bc3ccb91faac7e72c97041583352e
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/cea-hpc/rbh-fsevents/+/554991
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>