rbh-find-lustre: fix FID parsing
The fallback parsing of FID when llapi_fid_parse is not available
resulted in an invalid memory read when using an invalid FID of the
form:
rbh-lfind -fid [0x0:0x1:0x2

sscanf would fail to parse the format string "[" SFID "]%n". Since %n
was after ], sscanf would not initialise n_chars. This patch switches
%n and ]. But sscanf doesn't count %n in its return value. Therefore,
scanf will return 3 in this case and when the FID is correctly formed.

Therefore, the check for rc does not indicate whether there is a ] at
the end or not. We need to check the last caracter at index n_chars.

Change-Id: I8fb1522c52e5c4478862cb4da37a91532e5767d4
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>