rbh-find-lustre: fix conditional use of llapi_fid_parse
config.h was not included in filters.c which means that the presence of
llapi_fid_parse was not known by the compiler and the fallback option
was always used.

Since llapi_fid_parse allows any number of braces, a check for
consistency is added when llapi_fid_parse succeeds.

Change-Id: I224d32097476ced7670cab624d44ba8231c4e70d
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>