rbh-find-lustre: fix conditional use of llapi_fid_parse
config.h was not included in filters.c which means that the presence of
llapi_fid_parse was not known by the compiler and the fallback option
was always used.

Since llapi_fid_parse allows any number of braces, a check for
consistency is added when llapi_fid_parse succeeds.

Change-Id: I224d32097476ced7670cab624d44ba8231c4e70d
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/556479
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>
Tested-by: VALERI Yoann <yoann.valeri@cea.fr>