rbh-fsevents: implementation of the dedup logic
This first patch introduces a new type `struct rbh_fsevent_pool' which
will take in new events from the source and insert them in a hash table
indexed by the ID of the fsevent received.

If the ID is already in the map, the event will need to be deduplicated
if necessary. This patch does not do any deduplication yet but simply
inserts new events in a linked list.

Once enough events are received, the pool is considered full and a flush
is necessary otherwise, no new events can be inserted in it.

The flush will take the oldest entries in the pool until a flush
threshold is reached. Ideally, we should flush entries whose last events
are the oldest. This logic will be implemented in a later patch.

This patch also moves the check for errno from feed into the function
calling rbh_iter_next to get the final fsevent:
- file_sink_process when outputting YAML events
- mongo_bulk_init_from_fsevents when syncing to mongo

The reason for this is that previously, the iterator would only ever
yield one fsevent. But since now events are batched, the previous
behavior would skip a whole batch of fsevents.

Change-Id: Ifbf0230a77fabdb60b5effd9bcbd35ee64fbc8b3
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>