rbh-fsevents: dedup link/unlink
When we receive an unlink event, we look for a corresponding link
event in the list. The link event matches only if we have the same
id and the same parent in the link fsevent. If a match is found, we
can remove the link event from the list and ignore the incoming unlink
event.

Note that this patch does not deduplicate delete. Which means that the
last unlink cannot be deduplicated yet.

Change-Id: Icc2a4521aef5fc4a5816b56e6a7ac1e65aaabde7
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/557623
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>