rbh-fsevents: dedup delete events
Since we already deduplicate link/unlink events, we know when receiving
a delete, that we have at most one link event in the list.

If the link event is found, it means that all the create/unlink pairs
happened in the same batch of requests. Therefore, we can safely ignore
the delete, the entry was not yet inserted in the database. In this
case, all the events corresponding to this id are ignored and removed.

If the link event is not found, it means that the create event was not
in this batch, therefore the entry is probably stored in the database.
We need to keep the delete. All the other events can be ignored.

Change-Id: Ie8d0c2e8dccb378ab6811d17f2f0ec2d1778c1b0
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>