branching: fix unwritten path for the branch root
This patch fixes an issue where a sync done using a branch would result
in the branch entry in the database not having a path or name. This was
because the FTS entry didn't have a proper parent set, so the namespace
attributes wouldn't be considered for synchronization for that entry.

Change-Id: Ia1a7757356e98e73c7835f281b07e1d69ad9cef7
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>