cli: manage source argument as an URI
This patch modifies the source reading in the CLI from a path to a
file/lustre changelog MDT (with --lustre) option/'-' for stdin, to a
URI. The required scheme for this URI is "src", as it may not be a
proper RobinHood source.

Stdin is kept as-is, while the rest will have to be of the form
"src:backend_type:src_reader", with "backend_type" currently either
being "lustre" or "file", and "src_reader" the path of the file or
the MDT name containing the changelogs to read.

Also fix an issue where, if a given YAML file had an event with an
empty map, the parser would error out. This is because the parser is
waiting for a "start" event, but he receives an "end" event in that
case.

Change-Id: I11294e69bfdea15ebb9dd46c1b90de529bb3ae9d
Signed-off-by: Yoann Valeri <yoann.valeri@cea.fr>