file: handle yaml files more generically
This patch modifies the "file" source file to divide it into 2, one
specific to YAML files following the expected RobinHood fsevents
nomenclature (the new "file" source), and the other provides more
general functions and structures to read YAML files.

This is to prepare the Hestia backend, as the provided events will
be given in a YAML file, but not following standard fsevents
nomenclature.

Change-Id: I920a5c3d2d80e3f08ec7bb05b6659f39d4ef3096
Signed-off-by: Yoann Valeri <yoann.valeri@cea.fr>