rbh-fsevent: manage list elements' memory correctly
Since we are using rbh_sstacks for memory allocation in the
deduplicator, we cannot simply free elements allocated. One cannot free
a single element pushed in the stack apart from the last element pushed
(if we know its size).

This patch divides the allocations so that we can put list nodes in a
separate free list when they are no longer needed. On allocation, we
first look into the free list to see if we have nodes available. If so
we take one from the list and return it. Otherwise we have to allocate a
new one from the stack. This has the added benefit that once rbh-fsevent
has allocated enough nodes, no more allocations will be required which
means that long running rbh-fsevents instances will stop doing mallocs
after a few iterations.

Change-Id: I8411d6448348573c98d917f1d4b278dbfc134ce9
Signed-off-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1170932
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>