serialization: modify parser to properly handle sub-maps
This patch modifies the behaviour of the "parse_rbh_value_map" function
to take in a boolean signifying whether the function should do an
initial parsing or not.

The problem this solves is that, if you have the following map:
```
map-start:
  sub-map:
    key: value
```

And try to parse it, we would call "parse_rbh_value_map" on event start,
parse the token (which would be "map-start"), find it is a scalar
(corresponding to the string itself), parse the next token (which would
be "sub-map"), call "parse_rbh_value" on that token, detect it is the
start of a map, and call "parse_rbh_value_map" again.

Here, we would parse the next token, which is the scalar "key", and
since it isn't the start of a map, the function "parse_rbh_value_map"
would error out.

To prevent this, the function "parse_rbh_value_map" is changed to take
in a boolean. If that boolean is true, we do the exact same behaviour as
before. If not, we do not parse the next token directly, and instead
start parsing what we consider to be the content of the map.

Change-Id: I249f57f30eaabaad3e2a531cf7cc4bf94b55f656
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1170609
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>