serialization: modify parser to properly handle sub-maps
This patch modifies the behaviour of the "parse_rbh_value_map" function
to be able to take in an event, and start parsing from it instead of the
following one.

The problem this solves is that, if you have the following map:
```
map-start:
  sub-map:
    key: value
```

And try to parse it, we would call "parse_rbh_value_map" on event start,
parse the token (which would be "map-start"), find it is a scalar
(corresponding to the string itself), parse the next token (which would
be "sub-map"), call "parse_rbh_value" on that token, detect it is the
start of a map, and call "parse_rbh_value_map" again.

Here, we would parse the next token, which is the scalar "key", and
since it isn't the start of a map, the function "parse_rbh_value_map"
would error out.

To prevent this, the function "parse_rbh_value_map" is changed to take
in a pointer to an event. If that pointer is NULL, we do the exact same
behaviour as before. If it isn't NULL, we do not parse the next token
directly, and instead consider the given event as that start of the map.

Change-Id: I249f57f30eaabaad3e2a531cf7cc4bf94b55f656
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>