rbh-fsevent: manage fsevent deduplication seperately
Since we allocate memory for rbh_fsevents using an rbh_sstack, we cannot
free an individual rbh_fsevent when they are flushed to the database.

This patch moves the rbh_sstack from the rbh_fsevent_pool into the
rbh_fsevent_node structure. This means that we have one stack per
fsevent. But since we reuse those structures for incoming events this
extra cost will be amortized after a few iterations as long as the pool
size is bigger than the number of events we have to manage.

Change-Id: Id6e72da16935cdf01658a697fd0e6d5ea02a18e9
Signed-off-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1170933
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>