hestia: use time part of events as object time
This patch modifies the interpretation of the `time` part of the events
to use them as the object's time, depending of the event type.
CREATE events will use it as access, modify, change and birth time;
UPDATE events will use it as change time; REMOVE will keep ignoring it.

The next patch that will manage READ events will use it as access time.

Change-Id: I860e9052ecaa1298dfcbc56359da782299b545b6
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>