hestia: add main enrich function
This patch adds the two main enrich functions, the first one will check
if the "rbh-fsevents" map is present in the event's xattrs, and the
second one will enrich the event. Currently, the second one will only
retrieve all the Hestia attributes of the object, but that will be
changed.

Change-Id: I73a0efd930b9de825300a2399121c648b6d0858e
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1170710
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>