hestia: add enrichment tag when specific events occur
This patch modifies the behaviour of the Hestia source to create a new
event that will signify the enricher information should be retrieved. To
this new event, we add the enrichment tag "rbh-fsevents" and the key
"hestia". This event is generated when a CREATE or UPDATE happen.

Also ensure rbh-fsevents is built and installed even if the hestia
library is not found.

Change-Id: Ic9a21b08f5526b9b86a9d4b8668e5699fdca79c0
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>