tests: move lustre tests to specific directory
This patch moves all the Lustre tests to a new directory, and divide the
test_utils file in two: lustre_utils and test_utils. The first one
contains everything required for the Lustre tests, while the second
is/will be general to all tests.

This patch also modifies the run_tests function to take in sub setup
and teardown functions that will be dependant on each type of test
(Lustre, and in the next patches, Hestia).

Change-Id: Ic935f6f96d16d1767c17711faba16a97c27ac2ca
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1171056
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>