hestia: use iterator to return source events
This patch modifies the behaviour of the Hestia source to use an fsevent
iterator instead of storing the attributes of events that will be
generated during the next call to the source iter next.

This allows us to create a batch of fsevents per source events (i.e.
CREATE events will generated an iterator on 3 events), removing the need
to store additional attributes for a later call and simplifying the
behaviour of the source.

Currently, we only modify the behaviour of the two additionnal events
created in case of an UPDATE and CREATE, but these will be modified in
the next patch to properly initialize this iterator.

This modification will be ported later to the Lustre source aswell.

Change-Id: If2ab69ddcfd425e6ea2349c9066b1db5b32c0aa2
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1171423
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>