yaml: add void * sub item for different sources
This patch adds a void * field to the yaml source so that each can use a
field or structure specific to itself. For instance, the Hestia source
didn't use a rbh_fsevent but a rbh_iterator, while it was the contrary
for the File source. This field is held by the YAML source itself, and
should be freed by it.

Change-Id: I4d19ad8dcfb3f4786d34aa746ba7fb9cf7422957
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>