rbh-fsevents: move new events at the end of the IDs list
The global IDs list is used to choose which entries to flush when the
pool is full. The entries at the start of the list will be flushed.

Entries that just received an event are more likely to receive new ones
after a flush. To avoid flushing those entries, the global list is
reordered when a new event is received. The entries are pushed at the
end of the list. This means that entries are ordered from the oldest to
newest received event. So the flush function will start with the oldest
ones.

Change-Id: I047e2766f662d9bbf01658784e9609be55bf42fe
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>