hestia: use object name as inode name
This patch changes the name kept in the fsentry, from a simple
representation of the higher/lower 64 bits of an Hestia Id to the name
provided in the Hestia object. Since this name is optional and may not
be filled, we still keep the asprintf done to write the Hestia Id as a
string.

Change-Id: I6ad80d7d2f589d21f3be150500b82ac335383608
Signed-off-by: Yoann Valeri <yoann.valeri@cea.fr>