hestia: update retrieved times according to API changes
This patch modifies both the documentation to remove mention of the
changelog time not being the correct one, and the mtime being recorded.

The first is due to a modification in Hestia that makes the times
specified in each event correspond to the correct object time for that
kind of event (i.e. read = atime, update = ctime, create = all).

The second is because the information retrieved from the API only
contains access, change and birth time, so we stick to this behaviour in
the event feed aswell.

Change-Id: I15421c16fa57202aac5dfca89ea1b8e67541ccb3
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>