hestia: remove enricher setup
This patch removes the enrichment events created in the source. This can
be done because the Hestia event feed has been updated to contain
snapshots of the objects when an event occurs, so every information
about the object is already available in the events, and there is no
further need to interrogate the object store.

Also fix an issue with the order of update events, where the xattr
modification was enacted before the upsert. This is a problem because in
the Mongo backend, an xattr event will not create an entry, but an
upsert does. So the entry would be created after the xattr, and the
content of the xattr would be forgotten.

Change-Id: Ic439e152dfa5d13e3ea1bcb262b6353728d80a26
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1173709
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>