hestia: remove enricher setup
This patch removes the enrichment events created in the source. This can
be done because the Hestia event feed has been updated to contain
snapshots of the objects when an event occurs, so every information
about the object is already available in the events, and there is no
further need to interrogate the object store.

Change-Id: Ic439e152dfa5d13e3ea1bcb262b6353728d80a26
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>