librobinhood : Add support for /xxx mountpoint
The posix_iter_next function, when retrieving the
parent directory path, sets only the last "/" of
the path string to 0.

This leads to an issue where, if the mount point is at
the root directory, such as "/lfs01," the resulting
string will be "\0lfs01," causing the rbh-sync command
to report a "No such file" error.

To address this, a check is implemented to determine
if the mount point is at the root directory. If it is,
the operation to set the last "/" to 0 is skipped.