posix branch: use URI path as iterator root
This patch changes the behaviour of the branching in a POSIX backend, by
changing how the path used to iterate over each entry is transmitted
from the URI to the POSIX backend.

Before, when creating a new POSIX branch backend from an URI, we would
open the fsname in the URI, and use the resulting file descriptor
alongside the path of the branch to create a rbh_id (which required a
call to name_to_handle_at).

However, using this rbh_id, we would then create a branch backend, and
then when calling the POSIX branch backend's filter function, we would
convert that ID back to a path, by way of using an open_by_handle_at.
This double-conversion is inefficient, but may also be impossible if
some of these syscalls are forbidden on the system, which would make the
branch functionnality unusable.

To simplify and correct this issue, this patch removes this double ID
conversion if the full path is usable, by retrieving the current working
directory, adding the fsname and path to it, and giving the result to
the POSIX branch backend, removing the need for the filter function to
convert an ID back, instead simply using this full path provided.

Change-Id: I942de50d554efbb61ac9dd9daafe51b2e009a7eb
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>