librobinhood: use FTS info to avoid useless open
This patch changes the behaviour of the posix backend: instead of
opening an entry once and then a second time if it is a symbolic link,
we use the fact that the FTS API already gives us that information,
ensuring that in the case of symbolic links, only one open is necessary.

The one left issue is with sockets, as FTS cannot give us that
information, so they still require two opens.

Change-Id: Iacf142796f9c454c1467e3ab81bbc53f43df47ee
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>