rpm: add required files for IO-SEA specific RPM
This patch adds the specfile, global meson and utils files necessary for
generating a RPM. The specfile is specific to the IO-SEA project, so it
does not install any Lustre component.

Since it is specific to the IO-SEA project and it will end in a few
months, this specfile take some liberties with regard to how it manages
the repo in the RPM, including 2 repos that are not in the main
repository, for easier build. As such, it will not stay long in this
repo.

Change-Id: Ie14840aaabb5f11900d43a8a2ce0b2404fb9f3eb
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1176383
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>