posix: revert FTS change regarding symbolic links
This patch reverts the modification done to the POSIX entry opens in
Iacf142796f9c454c1467e3ab81bbc53f43df47ee.

This is necessary because when synchronizing a filesystem containing
symbolic links, the change actually wouldn't work because FTS cannot
give the info about whether the entry is a symbolic link or not.

However, it managed to do so when synchronizing a single entry, so the
case is rare enough to warrent reverting back the code to make it more
readable instead of just adding the ELOOP condition to the second if and
keeping the special single-entry-sync case.

Also this patch isn't a basic git revert because we want to keep the
test.

Change-Id: I83f9efb772270166f977aec01c666bba432ff236
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>