retention: remove relative and absolute expiration macro
This patch removes the two macro that set the expiration date in the
database to be either relative or absolute, because there is actually no
difference between them in the current version of the code.

When an absolute expiration date is set on an entry, the timestamp is
added to the DB. On the other hand, if it is a relative expiration date,
we calculate the actual date using this relative value and the maximum
between the access and modify time of the entry. When done, we set this
date at the expiration date in the DB.

So in both cases, what we have in the DB is a timestamp corresponding to
when the entry expires, meaning there is no difference between those two
modes.

Change-Id: I38733bd68aad8a6ad29042094a831895a3a6f9f3
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1176817
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>