retention: overload actions and printf for rbh-lfind
This patch adds an overload to the actions for rbh-find-lustre, and
especially the printf/fprintf actions. For this overload, we define the
'%E' directive, which will print the expiration date of an entry, if
available. For the sake of a smaller patch, this directive currently
only prints ENOTSUP, but it will be changed in the one.

Change-Id: I3e08ecb00604c7684eda56ff461e7ed9ad98df24
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>