lfind: create the '%E' directive
This patch adds the actual behaviour of the '%E' directive, which will
go through each entry and print their expiration date as set in the
inode xattrs. If they do not have one, or an error occurs, "None" is
printed instead. Also add the corresponding test.

Moreover, we change the way the "infinite" retention date is kept in the
database, from the max value of an unsigned 64 bit integer to a signed
64 bit integer. This is necessary because since Mongo can't handle
unsigned values, it is recorded in the database as -1, but when we
recover the value in rbh-find-lustre, the type given is as a signed
64 bit integer. So instead, it is simpler to keep the value as an
int64_t, which is big enough to contain an expiration date and will
avoid the conversion from int to uint.

Another way is what we chose to do here: in case the user wants to set
an infinite retention value, we use the max value of a int64_t, so Mongo
won't have any problem keeping that value and we won't have any problem
understanding it in rbh-find-lustre.

Change-Id: I50c250d5d27dccc2eb7c6fe65b68c5c852aac616
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1176820
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>