retention: keep the user xattrs in the database
This patch changes the behaviour of the retention mechanism to keep both
the expiration date and the expiration attribute set by the user in the
database.

Before this patch, we would fetch the expiration attribute, and then
change it to the expiration date, so the DB wouldn't see the attribute,
and the user wouldn't know we keep an actual timestamp instead of what
they set.

Change-Id: I8a88229333564b749b7b9a50f0345d107369da1f
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>