lfind: add '%I' directive
This patch adds the '%I' directive to print the rbh_id. This will be
useful for the retention script as we plan of creating a special event
which will be given to rbh-fsevents. For this we need the ID of the
entry to modify, which we don't have without this patch and without
doing a mongo request.

Change-Id: I4f32f3fe44a4eb2a360c00459aaa4f54f8a3ffaa
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>