utils: add retention script
This script will update every expired directory based on the most
recently accessed entry in them. The goal is to simplify the retention
mechanism, so that instead of assigning an expiration date to each and
every file in a directory, one can just set the expiration date to the
directory, and it will be deleted when expired.

The script works like this:
```
for every expired directory in a given URI:
    get the most recent accessed entry in it
    if the directory's expiration attribute + the entry's access date
    is superior to the directory's expiration date
        set the directory's expiration date to the sum of directory's
        expiration attribute and the entry's access date
    else
        delete the directory (not implemented yet)
```

Change-Id: I64e2ae6dee0b0e335f32f869f463686bc92e2818
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1176825
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>