lustre-mpi: create mpi_iterator to be used in lustre-mpi backend
Add the mpi iterator structure, it is inspired by the posix_iterator
structure. The difference is that it contains an mfu_flist instead of FTSENT* and FTS*. It also contains the current index in the mfu_flist and the size of
the mfu_flist.

Also create functions for creating, destroying and iterating the iterator. At
the creation, it create a mfu_flist by walking all the tree file. We make sure
we don't do a stat during the walk. This will be done later in the
luster_mpi_iterator_next function the same way as in the posix backend.

Finally, add the meson.build in src/backends/lustre-mpi for the compilation of
the future luster-mpi backend.

Change-Id: I8afddf01bab28b7fdd108028d5bb58f057b15fd8
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.ocre@cea.fr>