retention: update expiration date when using rbh-fsevents
This patch adds the management of the retention when using the
rbh-fsevents command. This is done by retrieving the xattr on the entry,
calculating the date, and then storing it in the fsevent.

This will ensure the expiration date is kept up to date with what
happens on the system and the different modifications to the files,
without requiring a new complete synchronization.

Change-Id: Iad5a8b1413451e2c6f93a5cde9e5948de2869e06
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1177105
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>