retention: keep retention attribute in fsevents
This patch ensures the attribute set by the retention attribute by the
user is set and kept in the database as a string. Since this attribute
can only be corrected understood by the Lustre backend (as POSIX will
only keep it as a binary value, and not a string), we must prevent the
POSIX enricher from enriching the retention attribute again, else we
would try to set the same attribute in the database twice, which would
provoke an error.

To do this, the POSIX enricher will go through each already retrieved
attributes, and check if the retention attribute was seen of not. If
yes, skip its retrieval from the POSIX enricher.

Change-Id: If31542cc934af136921835cdc760ee281bae63a3
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>