retention: keep retention attribute in fsevents
This patch converts the retention attribute set by the user to a string
in the database. Since this attribute is only interpreted by the Lustre
backend (as POSIX will only keep it as a binary value, and not a
string), we must prevent the POSIX enricher from enriching the retention
attribute again, otherwise we would try to set the same attribute in the
database twice, which would provoke an error.

To do this, the POSIX enricher will go through each already retrieved
attributes, and check if the retention attribute was seen of not. If
yes, skip its retrieval from the POSIX enricher.

Change-Id: If31542cc934af136921835cdc760ee281bae63a3
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1177106
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>