lustre-mpi: create a fsentry from mfu_flist entries
Modification of the creation of fsentry with the addition of the
fsentry_from_any function which allows you to construct a fsentry in a generic
way. Depending on the backend, we use the function fsentry_from_ftsent or
fsentry_from_mpi_fi which calls fsentry_from_any.

Also, add "mpi_file_info" structure which contains all the information
necessary to construct a fsentry from an entry in a mfu_flist. As an entry in
the mfu_flist is a path, we must reconstruct the name of the file/directory,
and determine its parent. To do this, we use the new functions "compute_name"
and "get_parent_id".

Finally, we need to recalculate the depth of the entry in the tree because if
our root is "foo/test" or "./test", mpifileutils gives that the depth of the
root is 1 instead of 0.

Change-Id: Ia67f9b71cf125770b411aff5cdaf5cf817b68bc5
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.ocre@cea.fr>