lustre-mpi: create a fsentry from mfu_flist entries
Modification of the creation of fsentry with the addition of the
fsentry_from_any function which allows you to construct an fsentry in
a generic way. Depending on the backend, we use the function
fsentry_from_ftsent or fsentry_from_mpi_fi which calls
fsentry_from_any.

Also, add "mpi_file_info" structure which contains all the information
necessary to construct an fsentry from an entry in an mfu_flist. As an
entry in the mfu_flist is a path, we need to determine its parent and
construct its id.

Change-Id: Ia67f9b71cf125770b411aff5cdaf5cf817b68bc5
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.ocre@cea.fr>