lustre-mpi: remove compute path for each entry
This patch add the parameter "root_depth" in the mpi_iterator structure. It
allows to avoid recalculating the depth for each entry in the mfu_flist. We
just need to subtract the depth of the root with the depth calculated by
mpifileutils. Useful in the case where the root given as input is "foo/../test"
or "./test". In this case, the depth will indeed be 0 instead of 1 and more as
mpi calculates it.

Change-Id: Ie5d3ed9d6907cedaa32df4d53a83134bc616af95
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.ocre@cea.fr>