lustre-mpi: add lustre_mpi backend and backend_filter
This patch add the new LUSTRE_MPI backend and a new backend_filter
function. It is based on the POSIX backend but overloads it like the
LUSTRE backend does. As it uses the mpi_iterator and not the
posix_iterator, the type of the iter_new function in the posix_backend
structure has changed from "struct posix_iterator *" to "void *". This
now allows you to use the POSIX backend with new iterators without
having to recreate a whole new backend.

Also, the unit test for this backend is almost the same as for the
LUSTRE backend because we want to have the same behavior with
mpifileutils. The only change is for the test with an empty root. In
mpifileutils, the first process always starts the walk from the root.
If the root is empty, only the first process flist is not empty. So,
for one process, the test is exactly the same as for the LUSTRE
backend and for the others processes, we need to check that the
iterator return NULL and set errno to ENODATA.

Change-Id: I96c359c09a69b23ff8b67b5df61bf27fbc94fdd3
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.ocre@cea.fr>