rbh-fsevent: correcting the error of not removing the node from free_ids after obtaining it.
When using the command 'rbh-fsevents -f 1 src:lustre:lfs01-MDT0000 -'
for synchronization, rbh-fsevents reports a 'Segmentation fault (core dumped)' error.
The reason is that each time a node is retrieved from free_ids, it is not removed from
the list, causing each subsequent retrieval from free_ids to be the same node.
Therefore, the code logic is modified to remove the node from free_ids if it is obtained
from there.