rbh-sync: add lustre-mpi backend
This patch adds the LUSTRE-MPI backend in rbh-sync. MPI and mfu are
initialized if there is "-mpi" in the first URI's backend name.
The function "sync" in rbh-sync is renamed "_sync" because "sync" is
already declared in <unistd.h> which is imported by "mfu.h".

Also, this patch adds the possibility to use special characters in
URI's backend names such as "rbh:lustre-mpi:.". Special characters are
replaced by "_" to match the definition of _RBH_BACKEND_PLUGIN.

Change-Id: I49415f506a2ad628d9ca40cf6df6a3fd8da3a9ec
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.ocre@cea.fr>