rbh-sync: add lustre-mpi script
This script uses the same tests as for the LUSTRE backend but with
MPI, to test if the behavior is exactly the same as without MPI. The
test functions are not anymore in the script but call from
"test_lustre_utils.bash".

Also, the two Lustre's scripts export now a variable to tell if we
want to use MPI or not while synchronizing. Finally, add a new
function in "test_utils.bash" to run rbh-sync with mpirun.

Change-Id: Icf4969413a235227120a3b797def60664efa0f98
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.ocre@cea.fr>